Rename partition to batch to avoid subtle bugs where previous partition method was used, now window. Fix method declarations in DoubleSequence using double as parameters where longs should be used in step, limit, skip, count. Renamed startingAt to increasingFrom and added decreasingFrom to primitive sequences. Added positiveFromZero and negativeFromZero to Int, Long sequences. Added range with step method to primitive sequences. Added accuracy parameter to methods in double sequence testing for termination, until, endingAt, range. Added steppingFrom method to primitive sequences.

+++ src/main/java/org/d2ab/iterator/doubles/LimitingDoubleIterator.java
@@ -19,10 +19,10 @@ package org.d2ab.iterator.doubles;
 import java.util.NoSuchElementException;
 
 public class LimitingDoubleIterator extends UnaryDoubleIterator {
-	private final double limit;
+	private final long limit;
 	private double count;
 
-	public LimitingDoubleIterator(DoubleIterator iterator, double limit) {
+	public LimitingDoubleIterator(DoubleIterator iterator, long limit) {
 		super(iterator);
 		this.limit = limit;
 	}
