+++ src/main/java/org/d2ab/iterator/doubles/SkippingDoubleIterator.java
@@ -19,11 +19,11 @@ package org.d2ab.iterator.doubles;
 import java.util.NoSuchElementException;
 
 public class SkippingDoubleIterator extends UnaryDoubleIterator {
-	private final double skip;
+	private final long skip;
 
 	private boolean skipped;
 
-	public SkippingDoubleIterator(DoubleIterator iterator, double skip) {
+	public SkippingDoubleIterator(DoubleIterator iterator, long skip) {
 		super(iterator);
 		this.skip = skip;
 	}
